1 添加微服务注册新框架Nacos  2 使用Fengin在Edu-Service中远程调用Vod-Sevice删除阿里云视频

+++ service/service_edu/src/main/java/com/atguigu/eduservice/service/impl/EduVideoServiceImpl.java
@@ -1,6 +1,7 @@
 package com.atguigu.eduservice.service.impl;
 
 import com.atguigu.commonutils.R;
+import com.atguigu.eduservice.client.VodClient;
 import com.atguigu.eduservice.entity.EduVideo;
 import com.atguigu.eduservice.mapper.EduVideoMapper;
 import com.atguigu.eduservice.service.EduVideoService;
@@ -12,6 +13,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
 
+import javax.annotation.Resource;
 import java.util.List;
 
 /**
@@ -25,20 +27,24 @@ import java.util.List;
 @Service
 @Transactional(rollbackFor = Exception.class)
 public class EduVideoServiceImpl extends ServiceImpl<EduVideoMapper, EduVideo> implements EduVideoService {
+    @Resource
+    private VodClient vodClient;
 
     @Override
     public R remVidAndAliyu(String id) {
-
         LambdaQueryWrapper<EduVideo> laEduVideo = new QueryWrapper<EduVideo>().lambda();
         laEduVideo.eq(EduVideo::getId, id).select(EduVideo::getVideoSourceId);
         EduVideo eduVideo = baseMapper.selectOne(laEduVideo);
         if (StringUtils.isEmpty(eduVideo))
             throw new GuLiException(20001, "该章节下没有改小节课程:(");
+        // 删除阿里云视频
+        String videoSourceId = eduVideo.getVideoSourceId();
+        R r = vodClient.deleteVideo(videoSourceId);
+        if (!r.isSuccess())
+            throw new GuLiException(20001, "删除阿里云视频失败:(");
         int reslutDele = baseMapper.deleteById(id);
         if (reslutDele <= 0)
             throw new GuLiException(20001, "删除小节课程失败:(");
-        //TODO 删除阿里云视频
-        String videoSourceId = eduVideo.getVideoSourceId();
         return R.ok();
     }
 }
